split into separate files, add docs, add derives
